requirements.txt
transformers
torch
inference-gpu[yolo-world]>=0.9.12
pymupdf
pdf2image
python-dotenv
pillow
numpy
onnxruntime-gpu
comtypes
opencv-python
paddlepaddle-gpu
paddleocr
langchain
langchain-community
python-pptx
python-docx
scipy
einops
rapidocr_onnxruntime
google-genai
timm
pandas
tabulate
faster-whisper
google-api-python-client
fastapi
uvicorn
chromadb
streamlit
langchain-text-splitters
langgraph
youtube_transcript_api
firebase-admin
langchain-google-genai
httpx